refactor: split logic and presentational component
